Revert "Revert "Bjorncs/convert state ressources to request handler""
